Updated to include decimals in conversion.

+++ CurrencyCalculator/Calculator.cs
@@ -2,35 +2,35 @@
 {
     public class Calculator
     {
-        public decimal CalculateFromEurosToPounds(int euro)
+        public decimal CalculateFromEurosToPounds(decimal euro)
         {
             decimal exchangeRate = 0.85m;
             decimal result = ExchangeRate(euro, exchangeRate);
             return result;
         }
 
-        public decimal CalculateFromUsDollarsToPounds(int dollars)
+        public decimal CalculateFromUsDollarsToPounds(decimal dollars)
         {
             decimal exchangeRate = 0.82m;
             decimal result = ExchangeRate(dollars, exchangeRate);
             return result;
         }
 
-        public decimal CalculateFromPoundsToEuros(int pound)
+        public decimal CalculateFromPoundsToEuros(decimal pound)
         {
             decimal exchangeRate = 1.17m;
             decimal result = ExchangeRate(pound, exchangeRate);
             return result;
         }
 
-        public decimal CalculateFromPoundsToUsDollars(int dollar)
+        public decimal CalculateFromPoundsToUsDollars(decimal dollar)
         {
             decimal exchangeRate = 1.22m;
             decimal result = ExchangeRate(dollar, exchangeRate);
             return result;
         }
 
-        private decimal ExchangeRate(int currencyValue, decimal exchangeRate)
+        private decimal ExchangeRate(decimal currencyValue, decimal exchangeRate)
         {
             return currencyValue * exchangeRate;
         }
